Move Session entity to its own persistence context
