+++ src/main/java/org/icatproject/exposed/ICATRest.java
@@ -76,6 +76,7 @@ import org.icatproject.core.manager.EntityInfoHandler;
 import org.icatproject.core.manager.GateKeeper;
 import org.icatproject.core.manager.Porter;
 import org.icatproject.core.manager.PropertyHandler;
+import org.icatproject.core.manager.SessionManager;
 import org.icatproject.core.manager.PropertyHandler.ExtendedAuthenticator;
 import org.icatproject.core.manager.search.FacetDimension;
 import org.icatproject.core.manager.search.FacetLabel;
@@ -113,6 +114,9 @@ public class ICATRest {
 	@EJB
 	PropertyHandler propertyHandler;
 
+	@EJB
+	SessionManager sessionManager;
+
 	private Set<String> rootUserNames;
 
 	private int maxEntities;
@@ -122,7 +126,7 @@ public class ICATRest {
 	private Map<String, String> cluster;
 
 	private void checkRoot(String sessionId) throws IcatException {
-		String userId = beanManager.getUserName(sessionId);
+		String userId = sessionManager.getUserName(sessionId);
 		if (!rootUserNames.contains(userId)) {
 			throw new IcatException(IcatExceptionType.INSUFFICIENT_PRIVILEGES, "user must be in rootUserNames");
 		}
@@ -169,7 +173,7 @@ public class ICATRest {
 	public String write(@Context HttpServletRequest request, @FormParam("sessionId") String sessionId,
 			@FormParam("entities") String json) throws IcatException {
 
-		String userName = beanManager.getUserName(sessionId);
+		String userName = sessionManager.getUserName(sessionId);
 
 		List<Long> beanIds = beanManager.write(userName, json, request.getRemoteAddr());
 
@@ -216,7 +220,7 @@ public class ICATRest {
 			@FormParam("name") String name, @FormParam("id") long id, @FormParam("keys") String keys)
 			throws IcatException {
 
-		String userName = beanManager.getUserName(sessionId);
+		String userName = sessionManager.getUserName(sessionId);
 
 		long beanId = beanManager.cloneEntity(userName, name, id, keys, request.getRemoteAddr());
 
@@ -267,7 +271,7 @@ public class ICATRest {
 		} catch (JsonException e) {
 			throw new IcatException(IcatExceptionType.BAD_PARAMETER, e.getMessage() + " in json " + json);
 		}
-		String userName = beanManager.getUserName(sessionId);
+		String userName = sessionManager.getUserName(sessionId);
 		beanManager.delete(userName, beans, request.getRemoteAddr());
 	}
 
@@ -342,7 +346,7 @@ public class ICATRest {
 	@Path("port")
 	@Produces(MediaType.TEXT_PLAIN)
 	public Response exportData(@QueryParam("json") String jsonString) throws IcatException {
-		return porter.exportData(jsonString, entityManager);
+		return porter.exportData(jsonString);
 	}
 
 	/**
@@ -479,8 +483,8 @@ public class ICATRest {
 	@Produces(MediaType.APPLICATION_JSON)
 	public String getSession(@PathParam("sessionId") String sessionId) throws IcatException {
 
-		String userName = beanManager.getUserName(sessionId);
-		double remainingMinutes = beanManager.getRemainingMinutes(sessionId);
+		String userName = sessionManager.getUserName(sessionId);
+		double remainingMinutes = sessionManager.getRemainingMinutes(sessionId);
 
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 		JsonGenerator gen = Json.createGenerator(baos);
@@ -507,7 +511,7 @@ public class ICATRest {
 	public String isLoggedIn1(@PathParam("userName") String userName) {
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 		try (JsonGenerator gen = Json.createGenerator(baos)) {
-			gen.writeStartObject().write("loggedIn", beanManager.isLoggedIn(userName)).writeEnd();
+			gen.writeStartObject().write("loggedIn", sessionManager.isLoggedIn(userName)).writeEnd();
 		}
 		return baos.toString();
 	}
@@ -561,7 +565,7 @@ public class ICATRest {
 	public String isLoggedIn2(@PathParam("mnemonic") String mnemonic, @PathParam("userName") String userName) {
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 		try (JsonGenerator gen = Json.createGenerator(baos)) {
-			gen.writeStartObject().write("loggedIn", beanManager.isLoggedIn(mnemonic + "/" + userName))
+			gen.writeStartObject().write("loggedIn", sessionManager.isLoggedIn(mnemonic + "/" + userName))
 					.writeEnd();
 		}
 		return baos.toString();
@@ -876,7 +880,7 @@ public class ICATRest {
 		logger.debug("Using " + plugin + " to authenticate");
 
 		String userName = authenticator.authenticate(credentials, request.getRemoteAddr()).getUserName();
-		String sessionId = beanManager.login(userName, lifetimeMinutes, request.getRemoteAddr());
+		String sessionId = sessionManager.login(userName, lifetimeMinutes, request.getRemoteAddr());
 
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 		JsonGenerator gen = Json.createGenerator(baos);
@@ -902,7 +906,7 @@ public class ICATRest {
 	@Path("session/{sessionId}")
 	public void logout(@Context HttpServletRequest request, @PathParam("sessionId") String sessionId)
 			throws IcatException {
-		beanManager.logout(sessionId, request.getRemoteAddr());
+		sessionManager.logout(sessionId, request.getRemoteAddr());
 	}
 
 	/**
@@ -1019,7 +1023,7 @@ public class ICATRest {
 		if (query == null) {
 			throw new IcatException(IcatExceptionType.BAD_PARAMETER, "query is not set");
 		}
-		String userName = beanManager.getUserName(sessionId);
+		String userName = sessionManager.getUserName(sessionId);
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 		try (JsonReader jr = Json.createReader(new ByteArrayInputStream(query.getBytes()))) {
 			JsonObject jo = jr.readObject();
@@ -1213,7 +1217,7 @@ public class ICATRest {
 		if (maxCount == 0) {
 			maxCount = 100;
 		}
-		String userName = beanManager.getUserName(sessionId);
+		String userName = sessionManager.getUserName(sessionId);
 		JsonValue searchAfterValue = null;
 		if (searchAfter != null && searchAfter.length() > 0) {
 			try (JsonReader jr = Json.createReader(new StringReader(searchAfter))) {
@@ -1583,7 +1587,7 @@ public class ICATRest {
 	@Path("session/{sessionId}")
 	public void refresh(@Context HttpServletRequest request, @PathParam("sessionId") String sessionId)
 			throws IcatException {
-		beanManager.refresh(sessionId, lifetimeMinutes, request.getRemoteAddr());
+		sessionManager.refresh(sessionId, lifetimeMinutes, request.getRemoteAddr());
 	}
 
 	/**
@@ -1634,7 +1638,7 @@ public class ICATRest {
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 		JsonGenerator gen = Json.createGenerator(baos);
 
-		String userName = beanManager.getUserName(sessionId);
+		String userName = sessionManager.getUserName(sessionId);
 		if (id == null) {
 			gen.writeStartArray();
 			for (Object result : beanManager.search(userName, query, request.getRemoteAddr())) {
